Proprify a little more, aim is to make coo::operator coo explicit, 3

+++ 2D/Ising-bis.cpp
@@ -44,8 +44,8 @@ public:
             shift   = I;
             pattern = {{.5, .5}, {-.5, .5}, {-.5, -.5}, {.5, -.5}};
         }
-        for (size_t i = 0; i < w(); ++i)
-            for (size_t j = 0; j < h(); ++j) {
+        for (size_t i = 0; i < size_t(w()); ++i)
+            for (size_t j = 0; j < size_t(h()); ++j) {
                 Color c = to_Color(at({i, j}));
                 if (c == to_Color(0)) continue;
                 vector<cpx> p;
