+++ 2D/PercSEP.cpp
@@ -21,8 +21,8 @@ public:
                 if (prng.bernoulli(H['p'])) put({i, j}, prng.bernoulli(H['l']) ? 2 : 1);
     }
     void clean() {
-        for (size_t x = 0; x < w(); ++x)
-            for (size_t y = 0; y < h(); ++y)
+        for (size_t x = 0; x < size_t(w()); ++x)
+            for (size_t y = 0; y < size_t(h()); ++y)
                 if (at({x, y}) > 0) at({x, y}) = at({x, y}) + 100;
         bool dirty = true;
         while (dirty) {
